Added cfire. Most of the codebase there is unchanged, should work without issue.
Also added the function tcs.ui.RefreshEnabled and a very simple
commandline interface for AutoNav in case someone finds it useful to
make autonav on/off aliases and bind them somewhere.
tcs.ui.RefreshEnabled is to refresh the checks indicating a plugin is
enabled in case the plugin is disabled or enabled through a means other
than the TCS Config window. It must be called by plugins as needed.