More fine-grained controls for NPCs. Provides ability to select and set Hive/Station Guards/Strike Forces/Conquerable Assets/Other NPCs to be processed 'As Player', hostile, or friendly.
When Conquerable Assets are processed 'As Player', MakeFriends checks to see if you have an Owner key, or a User Key with IFF access to the current conquerable sector. If so, turrets are green, else red. Results are cached so as not to be doing string matches for every turret in radar range at least once a frame.

Also, tcs.BatchAddTitledControls has been modified and renamed tcs.BatchAddControls to reflect its greater versatility. The batch format and function itself have been changed to be significantly simpler and more robust.

Signed-off-by: Tyler Gibbons <calaihm@gmail.com>