Something like MAJOR changes to cfire. Currently in the middle of rewrite to support: - Fully programmable chainfire groups - Auto-configuring for simple setup -- i.e., hiding all the power if you don't want/need it by setting up weapons groups, optimal delay times, and cfire pattern for you if you let it. - UI that is clear and easy to use for users who only want basic, minimal functionality in addition to supporting highly advanced users with (possible) domain-specific language/mini-lua interface. Also, a definite, is a UI that will allow graphical setup of complex cfire chains in an intuitive way - (Possible) Ability to share cfire binds with each other in game via chat (may spread to rest of TCS settings)
Obviously not all this is done. The first two top-level bullets are nearing completion, but the third is barely even in the design stage at this point. As of now, this module is currently DUBIOUSLY FUNCTIONAL, so TCS does not load it in. Instead, it will load the chain_fire_old folder instead.

This update WILL NOT BREAK you current TCS installation if you choose to download it, it'll just give you extra bits to look/poke at.

Signed-off-by: Tyler Gibbons <calaihm@gmail.com>